MOD: modified dpos send transaction validation logic
