+++ internal/ethapi/const.go
@@ -1,0 +1,17 @@
+// Copyright 2019 DxChain, All rights reserved.
+// Use of this source code is governed by an Apache
+// License 2.0 that can be found in the LICENSE file
+
+package ethapi
+
+const (
+
+	// MaxVoteCount is the max number of voted candidates in a vot tx
+	MaxVoteCount = 30
+
+	// StorageContractTxGas defines the default gas for storage contract tx
+	StorageContractTxGas = 90000
+
+	// DposTxGas defines the default gas for dpos tx
+	DposTxGas = 1000000
+)
